feat: Flip output to landscape orientation
This commit updates the popup.html file to include a new CSS style that flips the output to landscape orientation. The text is rotated by 90 degrees and is displayed in portrait orientation inside the popup window. This makes it easier to read the extracted kanji, which are displayed as a single line of text.