refactor(popup): Change output format to list of tuples
Changed the output format of the `generateOutput` function in `popup.js` to be a list of tuples instead of a dictionary of arrays. The new format is easier to read and use. Updated the `generateOutput` function to create a flattened list of tuples using the `flatMap` method, and displayed the output as a string using `JSON.stringify`. No changes were made to other files.